Update various stuff

+++ src/utils.rs
@@ -1,5 +1,5 @@
 use approx::AbsDiff;
-use sfcgal_sys::{sfcgal_geometry_t, size_t};
+use sfcgal_sys::sfcgal_geometry_t;
 use std::{ffi::CStr, os::raw::c_char};
 
 use crate::errors::get_last_error;
@@ -40,8 +40,8 @@ pub(crate) fn check_computed_value(val: f64) -> Result<f64> {
 // (as it seems to not always end with a null byte)
 // from the pointer to uninitialized memory with give
 // to it earlier.
-pub(crate) fn _c_string_with_size(raw_ptr: *const c_char, size: size_t) -> String {
-    let slice: &[u8] = unsafe { std::slice::from_raw_parts(raw_ptr as *const u8, size as usize) };
+pub(crate) fn _c_string_with_size(raw_ptr: *const c_char, size: usize) -> String {
+    let slice: &[u8] = unsafe { std::slice::from_raw_parts(raw_ptr as *const u8, size) };
     let res = std::str::from_utf8(slice).unwrap().to_string();
     unsafe { libc::free(raw_ptr as *mut libc::c_void) };
     res
